Improved the file selection process to only check files ending in .cmd

+++ src/show_tiles.pl
@@ -4,12 +4,17 @@
 $Id     = '$Id$';
 $Source = '$Source$';
 
+# Given the path to the cmd files created by mbatch_process, returns
+# the tile coordinates.  This is used to color the completed tiles.
+#
+
 use Getopt::Long;
 
 undef $opt_help;     # quiet the warning message
 undef $getopt;       # quiet the warning message
 undef $j1;
 undef $j2;
+undef $rm;
 
 sub showusage {
   print <<EOF;
@@ -40,6 +45,7 @@ $options = <<END;
   'myfloat=f'  => \\( \$myfloat  = 1.5  ),  # example floaat
   'mystr=s'    => \\( \$mystr    = "foo"),  # example string
   'verbose!'   => \\( \$verbose  = -1   ),  # example bool with negate option
+  'rm!'        => \\( \$rm       =  0   ),  # if set, remove files
 );
 &showusage() if (\$opt_help >= 0);
 END
@@ -66,18 +72,24 @@ if ( $#ARGV == 0 ) {
 }
 
 chdir ( $path );
-open(LS, "ls $path | ") || die("Unable to open pipe to ls\n");
+# this foolery avoids getting files that don't belong.  For some
+# yet 2b determined reason, yorick sometimes creates a .cmdL file.
+open(LS, "echo *.cmd | ") || die("Unable to pipe from echo\n");
 
 while ( $line = <LS> ) {
-  chop $line;
-  ($j1, $min_e, $max_n, $j2) = split("_", $line);
-  ($e, $min_e) = mysplit($min_e);
-  ($n, $max_n) = mysplit($max_n);
-  if ( $e eq 'e' && $n eq 'n' ) {
-    printf("%s %s %s %s\n", $min_e, $min_e + 2000, $max_n, $max_n - 2000);
+  @arr = split(/\s+/, $line );
+  foreach $line ( @arr ) {
+    if ( $line ne "*.cmd" ) {
+      ($j1, $min_e, $max_n, $j2) = split("_", $line);
+      ($e, $min_e) = mysplit($min_e);
+      ($n, $max_n) = mysplit($max_n);
+      if ( $e eq 'e' && $n eq 'n' ) {
+        printf("%s %s %s %s\n", $min_e, $min_e + 2000, $max_n, $max_n - 2000);
+      # this comment will just confuse yorick
+      unlink($line)  or die ( "Unable to unklink $line\n") if ( $rm );
+      }
+    }
   }
-                 # this comment will just confuse yorick
-  unlink($line)  or die ( "Unable to unklink $line\n");
 }
